Fixes Location auto tests for mobile platforms.
There was some incorrect behaviour where qreal was expected to
be double, which is true on the desktop but not on mobile platforms.

This change also removes some uses of the expected fail test macro
which were being used as the equivalent of QCOMPARE(!<expression>)

+++ tests/auto/qgeoboundingcircle/tst_qgeoboundingcircle.cpp
@@ -78,22 +78,22 @@ void tst_QGeoBoundingCircle::defaultConstructor()
 {
     QGeoBoundingCircle c;
     QVERIFY(!c.center().isValid());
-    QCOMPARE(c.radius(), -1.0);
+    QCOMPARE(c.radius(), qreal(-1.0));
 }
 
 void tst_QGeoBoundingCircle::centerRadiusConstructor()
 {
-    QGeoBoundingCircle c(QGeoCoordinate(1,1), 50);
+    QGeoBoundingCircle c(QGeoCoordinate(1,1), qreal(50.0));
     QCOMPARE(c.center(), QGeoCoordinate(1,1));
-    QCOMPARE(c.radius(), 50.0);
+    QCOMPARE(c.radius(), qreal(50.0));
 }
 
 void tst_QGeoBoundingCircle::comparison()
 {
-    QGeoBoundingCircle c1(QGeoCoordinate(1,1), 50);
-    QGeoBoundingCircle c2(QGeoCoordinate(1,1), 50);
-    QGeoBoundingCircle c3(QGeoCoordinate(1,1), 35);
-    QGeoBoundingCircle c4(QGeoCoordinate(1,2), 50);
+    QGeoBoundingCircle c1(QGeoCoordinate(1,1), qreal(50.0));
+    QGeoBoundingCircle c2(QGeoCoordinate(1,1), qreal(50.0));
+    QGeoBoundingCircle c3(QGeoCoordinate(1,1), qreal(35.0));
+    QGeoBoundingCircle c4(QGeoCoordinate(1,2), qreal(50.0));
 
     QVERIFY(c1 == c2);
     QVERIFY(c1 != c3);
@@ -111,9 +111,9 @@ void tst_QGeoBoundingCircle::radius()
 {
     QGeoBoundingCircle c;
     c.setRadius(1.0);
-    QCOMPARE(c.radius(), 1.0);
+    QCOMPARE(c.radius(), qreal(1.0));
     c.setRadius(5.0);
-    QCOMPARE(c.radius(), 5.0);
+    QCOMPARE(c.radius(), qreal(5.0));
 }
 
 void tst_QGeoBoundingCircle::center()
@@ -129,15 +129,15 @@ void tst_QGeoBoundingCircle::translate_data()
 {
     QTest::addColumn<QGeoCoordinate>("center");
     QTest::addColumn<qreal>("radius");
-    QTest::addColumn<qreal>("lat");
-    QTest::addColumn<qreal>("lon");
+    QTest::addColumn<double>("lat");
+    QTest::addColumn<double>("lon");
     QTest::addColumn<QGeoCoordinate>("newCenter");
 
-    QTest::newRow("from 0,0") << QGeoCoordinate(0,0) << 10.0 <<
+    QTest::newRow("from 0,0") << QGeoCoordinate(0,0) << qreal(10.0) <<
                                  5.0 << 5.0 << QGeoCoordinate(5.0, 5.0);
-    QTest::newRow("across 0,0") << QGeoCoordinate(-2, -2) << 20.0 <<
+    QTest::newRow("across 0,0") << QGeoCoordinate(-2, -2) << qreal(20.0) <<
                                    5.0 << 5.0 << QGeoCoordinate(3.0, 3.0);
-    QTest::newRow("backwards across 0,0") << QGeoCoordinate(5,5) << 50.0
+    QTest::newRow("backwards across 0,0") << QGeoCoordinate(5,5) << qreal(50.0)
                                               << -13.0 << 5.0
                                               << QGeoCoordinate(-8.0, 10.0);
 }
@@ -146,8 +146,8 @@ void tst_QGeoBoundingCircle::translate()
 {
     QFETCH(QGeoCoordinate, center);
     QFETCH(qreal, radius);
-    QFETCH(qreal, lat);
-    QFETCH(qreal, lon);
+    QFETCH(double, lat);
+    QFETCH(double, lon);
     QFETCH(QGeoCoordinate, newCenter);
 
     QGeoBoundingCircle c(center, radius);
@@ -172,13 +172,13 @@ void tst_QGeoBoundingCircle::valid_data()
     QTest::addColumn<qreal>("radius");
     QTest::addColumn<bool>("valid");
 
-    QTest::newRow("default") << QGeoCoordinate() << -1.0 << false;
-    QTest::newRow("empty coord") << QGeoCoordinate() << 5.0 << false;
-    QTest::newRow("NaN coord") << QGeoCoordinate(500, 500) << 5.0 << false;
-    QTest::newRow("bad radius") << QGeoCoordinate(10, 10) << -5.0 << false;
-    QTest::newRow("NaN radius") << QGeoCoordinate(10, 10) << qQNaN() << false;
-    QTest::newRow("zero radius") << QGeoCoordinate(10, 10) << 0.0 << true;
-    QTest::newRow("good") << QGeoCoordinate(10, 10) << 5.0 << true;
+    QTest::newRow("default") << QGeoCoordinate() << qreal(-1.0) << false;
+    QTest::newRow("empty coord") << QGeoCoordinate() << qreal(5.0) << false;
+    QTest::newRow("NaN coord") << QGeoCoordinate(500, 500) << qreal(5.0) << false;
+    QTest::newRow("bad radius") << QGeoCoordinate(10, 10) << qreal(-5.0) << false;
+    QTest::newRow("NaN radius") << QGeoCoordinate(10, 10) << qreal(qQNaN()) << false;
+    QTest::newRow("zero radius") << QGeoCoordinate(10, 10) << qreal(0.0) << true;
+    QTest::newRow("good") << QGeoCoordinate(10, 10) << qreal(5.0) << true;
 }
 
 void tst_QGeoBoundingCircle::valid()
@@ -197,13 +197,13 @@ void tst_QGeoBoundingCircle::empty_data()
     QTest::addColumn<qreal>("radius");
     QTest::addColumn<bool>("empty");
 
-    QTest::newRow("default") << QGeoCoordinate() << -1.0 << true;
-    QTest::newRow("empty coord") << QGeoCoordinate() << 5.0 << true;
-    QTest::newRow("NaN coord") << QGeoCoordinate(500, 500) << 5.0 << true;
-    QTest::newRow("bad radius") << QGeoCoordinate(10, 10) << -5.0 << true;
-    QTest::newRow("NaN radius") << QGeoCoordinate(10, 10) << qQNaN() << true;
-    QTest::newRow("zero radius") << QGeoCoordinate(10, 10) << 0.0 << true;
-    QTest::newRow("good") << QGeoCoordinate(10, 10) << 5.0 << false;
+    QTest::newRow("default") << QGeoCoordinate() << qreal(-1.0) << true;
+    QTest::newRow("empty coord") << QGeoCoordinate() << qreal(5.0) << true;
+    QTest::newRow("NaN coord") << QGeoCoordinate(500, 500) << qreal(5.0) << true;
+    QTest::newRow("bad radius") << QGeoCoordinate(10, 10) << qreal(-5.0) << true;
+    QTest::newRow("NaN radius") << QGeoCoordinate(10, 10) << qreal(qQNaN()) << true;
+    QTest::newRow("zero radius") << QGeoCoordinate(10, 10) << qreal(0.0) << true;
+    QTest::newRow("good") << QGeoCoordinate(10, 10) << qreal(5.0) << false;
 }
 
 void tst_QGeoBoundingCircle::empty()
@@ -223,15 +223,15 @@ void tst_QGeoBoundingCircle::contains_data()
     QTest::addColumn<QGeoCoordinate>("probe");
     QTest::addColumn<bool>("result");
 
-    QTest::newRow("own centre") << QGeoCoordinate(1,1) << 100.0 <<
+    QTest::newRow("own centre") << QGeoCoordinate(1,1) << qreal(100.0) <<
                                    QGeoCoordinate(1,1) << true;
-    QTest::newRow("over the hills") << QGeoCoordinate(1,1) << 100.0 <<
+    QTest::newRow("over the hills") << QGeoCoordinate(1,1) << qreal(100.0) <<
                                        QGeoCoordinate(30, 40) << false;
-    QTest::newRow("at 0.5*radius") << QGeoCoordinate(1,1) << 100.0 <<
+    QTest::newRow("at 0.5*radius") << QGeoCoordinate(1,1) << qreal(100.0) <<
                                       QGeoCoordinate(1.00015374,1.00015274) << true;
-    QTest::newRow("at 0.99*radius") << QGeoCoordinate(1,1) << 100.0 <<
+    QTest::newRow("at 0.99*radius") << QGeoCoordinate(1,1) << qreal(100.0) <<
                                        QGeoCoordinate(1.00077538, 0.99955527) << true;
-    QTest::newRow("at 1.01*radius") << QGeoCoordinate(1,1) << 100.0 <<
+    QTest::newRow("at 1.01*radius") << QGeoCoordinate(1,1) << qreal(100.0) <<
                                        QGeoCoordinate(1.00071413, 0.99943423) << false;
 }
